Store volume as float instead of int
Some tickers' daily volume exceeds PHP_INT_MAX on the 32-bit ARM
production host, silently corrupting an (int) cast during the first
production sync (observed as float-to-int cast warnings).

+++ src/Backtest/YahooFinanceClient.php
@@ -89,7 +89,7 @@ class YahooFinanceClient
                 low: (float) $quotes['low'][$i],
                 close: (float) $quotes['close'][$i],
                 adjClose: (float) $adjustedClose['adjclose'][$i],
-                volume: (int) $quotes['volume'][$i],
+                volume: (float) $quotes['volume'][$i],
             );
         }
